+++ app/renderer/modes.js
@@ -149,6 +149,9 @@ const setMode = m => {
     if (!modes[mode] || currentMode() === mode || !currentPage()) {
         return
     }
+    if (currentPage()?.isCrashed() && "fipv".includes(mode[0])) {
+        return
+    }
     if (modes[currentMode()].onLeave) {
         modes[currentMode()].onLeave(mode)
     }